requirements.txt
Django>=1.7,<1.8
pyScss==1.1.5
django-pagination==1.0.7
gunicorn==18.0
phonenumbers==6.0.0
django-emailtools==0.2.1
django-betterforms==0.1.3
django-localflavor==1.1
env-excavator==1.2.0
psycopg2==2.6
django-pipeline==1.4.6
raven==5.2.0
djangorestframework==3.0.5
boto==2.36.0
twilio==3.7.1
pylibmc==1.4.1
django-authtools==1.0.0
django-secure==1.0.1
django-cache-url==0.8.0
django-s3-folder-storage==0.2
